HUGE CHANGES: Merged both apps. Changed urls, views, models, settings, and template files to adapt to this change. MAKE SURE TO DELETE YOUR DATABASE FILE AND REMAKE IT USING MAKEMIGRATIONS
