+++ spunimusic/spunimusic/settings.py
@@ -41,7 +41,6 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'social_django',
     'spuni',
-    'userprofile',
 ]
 
 MIDDLEWARE = [
