+++ spunimusic/templates/index.html
@@ -1,4 +1,4 @@
-{% extends "userprofile/base.html" %}
+{% extends "base.html" %}
 {% load static %}
 
 {% block css_block %}
