+++ spunimusic/templates/song.html
@@ -1,6 +1,6 @@
 <!DOCTYPE html>
 
-{% extends 'userprofile/base.html'%}
+{% extends 'base.html'%}
 
 
 {% block body_block %}
