--- spunimusic/userprofile/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class UserprofileConfig(AppConfig):
-    name = 'userprofile'
